Update create_db.sql
student_id fix: small int wasn't enough to store actual unige student ids

+++ backend/server_db/sql_scripts/create_db.sql
@@ -5,7 +5,7 @@ CREATE DATABASE studybuds_db;
 \c studybuds_db
 
 CREATE TABLE Student (
-    student_id smallint PRIMARY KEY,
+    student_id int PRIMARY KEY,
     telegram_account int
 );
